chore(cli): move workflow helpers into dedicated boundary
