+++ src/cli/src/modules/dependencies.js
@@ -43,7 +43,7 @@ export {
     resetProgressBarRegistryForTesting
 } from "../runtime-options/progress-bar.js";
 
-export { ensureWorkflowPathsAllowed } from "../shared/workflow/path-filter.js";
+export { ensureWorkflowPathsAllowed } from "../workflow/path-filter.js";
 
 export { REPO_ROOT, resolveFromRepoRoot } from "../shared/workspace-paths.js";
 
@@ -51,7 +51,7 @@ export {
     DEFAULT_FIXTURE_DIRECTORIES,
     createPathFilter,
     normalizeFixtureRoots
-} from "../shared/workflow/fixture-roots.js";
+} from "../workflow/fixture-roots.js";
 
 export {
     writeFileArtifact,
