+++ src/cli/src/workflow/index.js
@@ -1,0 +1,2 @@
+export * from "./path-filter.js";
+export * from "./fixture-roots.js";
