+++ src/cli/test/manual-file-helpers.test.js
@@ -8,7 +8,7 @@ import {
     writeManualFile,
     writeManualJsonArtifact
 } from "../src/modules/manual/file-helpers.js";
-import { createWorkflowPathFilter } from "../src/shared/workflow/path-filter.js";
+import { createWorkflowPathFilter } from "../src/workflow/path-filter.js";
 
 function createTempDirFactory() {
     let counter = 0;
