+++ src/cli/test/performance-fixture-roots.test.js
@@ -5,7 +5,7 @@ import { describe, it } from "node:test";
 import {
     DEFAULT_FIXTURE_DIRECTORIES,
     normalizeFixtureRoots
-} from "../src/shared/workflow/fixture-roots.js";
+} from "../src/workflow/fixture-roots.js";
 
 describe("normalizeFixtureRoots", () => {
     it("includes default fixture directories by default", () => {
